pyproject.toml
[project]
name = "orka-reasoning"
version = "0.0.1"
description = "Modular agent orchestrator for reasoning pipelines"
authors = [{ name="Marco Somma", email="your@email.com" }]
dependencies = [
    "python-dotenv",
    "redis",
    "pyyaml",
    "openai",          # if you're using OpenAI
    "duckduckgo-search", # if you're using DDG search
]
classifiers = [
  "License :: Other/Proprietary License",
  "Programming Language :: Python :: 3"
]
dynamic = ["readme"]

[project.optional-dependencies]
dev = ["pytest", "coverage"]

[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project.urls]
"Homepage" = "https://github.com/marcosomma/orka"
"Repository" = "https://github.com/marcosomma/orka"
